rename namespace and factory method for extractor
